+++ webknossos/Changelog.md
@@ -37,6 +37,8 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 - Added `chunk_shape` and `shard_shape` arguments to `Layer.downsample`, `Layer.downsample_mag`, `Layer.redownsample`, `Layer.downsample_mag_list`, `Layer.upsample` and `RemoteLayer.downsample`. [#1434](https://github.com/scalableminds/webknossos-libs/pull/1434)
 
 ### Changed
+- The `webknossos convert` command now accepts a `--upload` flag, to upload the converted dataset to WEBKNOSSOS. In that case, the target path is not required anymore. [#1433](https://github.com/scalableminds/webknossos-libs/pull/1433)
+- The `webknossos downsample` and `webknossos upsample` commands now work with remote datasets on WEBKNOSSOS. [#1433](https://github.com/scalableminds/webknossos-libs/pull/1433)
 - The default shard shape for downsampled mags is now `(1024, 1024, 1024)` (default) or `(4096, 4096, 32)` (zarr3 and 32 sections or less) or `(32, 32, 32)` (zarr2) and the chunk shape is now `(32, 32, 32)`. [#1434](https://github.com/scalableminds/webknossos-libs/pull/1434)
 - The `axisOrder` field of the `MagViewProperties` is now always persisted in the `datasource-properties.json` file to avoid misinterpretation of the implicit axis ordering. [#1437](https://github.com/scalableminds/webknossos-libs/pull/1437)